Utils function file
