+++ .gitignore
@@ -2,6 +2,9 @@
 # API Keys
 api_key.js
 
+#DS_Store
+*.DS_Store
+
 # Logs
 logs
 *.log
