+++ react-client/src/index.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import $ from 'jquery';
 import Search from './components/Search.jsx';
+import Utils from './utils.js'
 
 class App extends React.Component {
   constructor(props) {
@@ -12,26 +13,8 @@ class App extends React.Component {
   }
 
   componentDidMount() {
-    let options = {
-      enableHighAccuracy: true,
-      timeout: 5000,
-      maximumAge: 0
-    };
-
-    let success = (pos) => {
-      var crd = pos.coords;
-
-      console.log('Your current position is:');
-      console.log(`Latitude : ${crd.latitude}`);
-      console.log(`Longitude: ${crd.longitude}`);
-      console.log(`More or less ${crd.accuracy} meters.`);
-    };
-
-    let error = (err) => {
-      console.warn(`ERROR(${err.code}): ${err.message}`);
-    };
-
-    navigator.geolocation.getCurrentPosition(success, error, options);
+    // ask user for location on page load
+    Utils.location();
   }
