+++ server/index.js
@@ -7,8 +7,8 @@ require('dotenv').config();
 console.log('port inside index', process.env.PORT);
 const port = process.env.PORT;
 
-app.use(bodyParser.json()); // for parsing application/json
-app.use(bodyParser.urlencoded({ extended: true })); // for parsing application/x-www-form-urlencoded
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
 
 app.use(express.static(__dirname + '/../react-client/dist'));
